refactor: Extract metric handle types into a dedicated header
This avoids transitive inclusion of the heavy prometheus headers.
