+++ src/storage/v2/constraints/existence_constraints.hpp
@@ -18,7 +18,7 @@
 
 #include "absl/container/flat_hash_map.h"
 
-#include "metrics/prometheus_metrics.hpp"
+#include "metrics/metric_handles.hpp"
 #include "metrics/scoped_gauge.hpp"
 #include "storage/v2/constraint_verification_info.hpp"
 #include "storage/v2/constraints/constraint_violation.hpp"
