+++ src/storage/v2/constraints/type_constraints.hpp
@@ -17,7 +17,7 @@
 
 #include "absl/container/flat_hash_map.h"
 
-#include "metrics/prometheus_metrics.hpp"
+#include "metrics/metric_handles.hpp"
 #include "metrics/scoped_gauge.hpp"
 #include "storage/v2/constraints/constraint_violation.hpp"
 #include "storage/v2/constraints/constraints_mvcc.hpp"
