+++ src/storage/v2/inmemory/unique_constraints.hpp
@@ -15,7 +15,7 @@
 #include <optional>
 #include <variant>
 #include "memory/db_arena_fwd.hpp"
-#include "metrics/prometheus_metrics.hpp"
+#include "metrics/metric_handles.hpp"
 #include "metrics/scoped_gauge.hpp"
 #include "storage/v2/constraints/active_constraints.hpp"
 #include "storage/v2/constraints/constraint_violation.hpp"
